Set ECDHE parameters on listening TLS sockets.
This configures settings for servers listening on TLS sockets so that
they can communicate with clients who want to use ECHDE ciphers. Without
this, we probably offer ECDHE ciphers to clients, but cannot actually
use them. Safari on iOS9 seems to have a preference for these ciphers.

ECDHE-RSA-AES256-GCM-SHA384 is an example of a cipher that this commit
allows the use of.